- no longer displaying "not logged in". This was too confusing for some
- fixed files missing from vfs if parent directory exists in real destination dir
- implemented plugin api to access current profile
- steam game detection now also works for 64-bit games
- removed reference to archive tab from tutorial
- usvfs log level is now taken from config
- some cleanup

+++ src/profile.h
@@ -210,7 +210,7 @@ public:
    * 
    * @return number of mods for which the profile has status information
    **/
-  unsigned int numMods() const { return m_ModStatus.size(); }
+  size_t numMods() const { return m_ModStatus.size(); }
 
   /**
    * @return the number of mods that can be enabled and where the priority can be modified
